FEC-4043 Changes in GoogleMediaFramework
Added MediaDrmCallback argument to createRendererBuilder().
Changed ExoplayerWrapper.seekTo() to get long as argument.
ExoplayerUtil: urlConnection.connect() before getOutputStream().

+++ googlemediaframework/src/main/java/com/google/android/libraries/mediaframework/exoplayerextensions/ExoplayerWrapper.java
@@ -639,7 +639,7 @@ public class ExoplayerWrapper implements ExoPlayer.Listener, ChunkSampleSource.E
    * Move the seek head to the given position.
    * @param positionMs A number of milliseconds after the start of the video.
    */
-  public void seekTo(int positionMs) {
+  public void seekTo(long positionMs) {
     player.seekTo(positionMs);
   }
